requirements.txt
tensorflow>=2.10.0
opencv-python>=4.5.0
numpy>=1.21.0
matplotlib>=3.5.0
scikit-learn>=1.0.0
scikit-image>=0.19.0
pyttsx3>=2.90
Pillow>=9.0.0
albumentations>=1.3.0
seaborn>=0.11.0
psutil>=5.8.0
SpeechRecognition>=3.10.0
pyaudio>=0.2.11
bcrypt>=4.1.3
pyotp>=2.9.0
cryptography>=42.0.8
python-dotenv>=1.0.1

python-pptx==0.6.21
